Use ReadyAI dataset for local training + updated readme
